pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cape-allocator"
version = "0.1.0"
description = "Component CAPE + Merton Rule equity/TIPS portfolio allocator"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "yfinance>=0.2",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "ty>=0.0.1",
    "hypothesis>=6.100",
]
api = [
    "fastapi>=0.100",
    "mangum>=0.17",
    "uvicorn>=0.23",
    "httpx>=0.25",
]

[project.scripts]
cape-allocator = "cape_allocator.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cape_allocator"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B", "A"]

[tool.ty.environment]
python-version = "3.11"

[tool.ty.rules]
# Treat all diagnostics as errors for strict checking
all = "error"

[tool.ty.analysis]
# Allow unresolved imports for missing or untyped dependencies
allowed-unresolved-imports = ["*.**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
pythonpath = ["."]

[tool.coverage.run]
source = ["cape_allocator"]
omit = ["*/cli/*"]
